+++ .eslintrc.js
@@ -13,6 +13,11 @@ module.exports = {
     {
       files: ['*.test.{ts,js}', '**/tests/**/*.{ts,js}'],
       extends: ['@metamask/eslint-config-jest'],
+      rules: {
+        // TODO: Re-enable
+        'import/no-named-as-default-member': 'off',
+        'jest/no-conditional-expect': 'off',
+      },
     },
     {
       // These files are test helpers, not tests. We still use the Jest ESLint
@@ -38,27 +43,41 @@ module.exports = {
     {
       files: ['*.ts'],
       extends: ['@metamask/eslint-config-typescript'],
+      parserOptions: {
+        tsconfigRootDir: __dirname,
+        project: ['./tsconfig.packages.json'],
+      },
       rules: {
         // disabled due to incompatibility with Record<string, unknown>
         // See https://github.com/Microsoft/TypeScript/issues/15300#issuecomment-702872440
         '@typescript-eslint/consistent-type-definitions': 'off',
 
-        // Modified to include the 'ignoreRestSiblings' option
-        // TODO: Migrate this rule change back into `@metamask/eslint-config`
-        '@typescript-eslint/no-unused-vars': [
-          'error',
-          {
-            vars: 'all',
-            args: 'all',
-            argsIgnorePattern: '[_]+',
-            ignoreRestSiblings: true,
-          },
-        ],
+        // TODO: auto-fix breaks stuff
+        '@typescript-eslint/promise-function-async': 'off',
+
+        // TODO: re-enble most of these rules
+        '@typescript-eslint/await-thenable': 'warn',
+        '@typescript-eslint/naming-convention': 'off',
+        '@typescript-eslint/no-floating-promises': 'warn',
+        '@typescript-eslint/no-for-in-array': 'warn',
+        '@typescript-eslint/no-loss-of-precision': 'warn',
+        '@typescript-eslint/no-misused-promises': 'warn',
+        '@typescript-eslint/no-unnecessary-type-assertion': 'off',
+        '@typescript-eslint/unbound-method': 'off',
+        '@typescript-eslint/prefer-enum-initializers': 'off',
+        '@typescript-eslint/prefer-nullish-coalescing': 'off',
+        '@typescript-eslint/prefer-optional-chain': 'off',
+        '@typescript-eslint/prefer-reduce-type-parameter': 'off',
+        '@typescript-eslint/restrict-plus-operands': 'warn',
+        '@typescript-eslint/restrict-template-expressions': 'warn',
+        'no-restricted-syntax': 'off',
+        'no-restricted-globals': 'off',
       },
     },
     {
       files: ['*.d.ts'],
       rules: {
+        '@typescript-eslint/naming-convention': 'warn',
         'import/unambiguous': 'off',
       },
     },
@@ -66,48 +85,38 @@ module.exports = {
       files: ['scripts/*.ts'],
       rules: {
         // All scripts will have shebangs.
-        'node/shebang': 'off',
+        'n/shebang': 'off',
       },
     },
   ],
   rules: {
-    // This is already set in the newest version of eslint-config
-    'padding-line-between-statements': [
-      'error',
-      {
-        blankLine: 'always',
-        prev: 'directive',
-        next: '*',
-      },
-      {
-        blankLine: 'any',
-        prev: 'directive',
-        next: 'directive',
-      },
-    ],
-
     // Left disabled because various properties throughough this repo are snake_case because the
     // names come from external sources or must comply with standards
     // e.g. `txreceipt_status`, `signTypedData_v4`, `token_id`
     camelcase: 'off',
+    'id-length': 'off',
 
     // TODO: re-enble most of these rules
+    '@typescript-eslint/naming-convention': 'off',
     'function-paren-newline': 'off',
     'guard-for-in': 'off',
+    'id-denylist': 'off',
     'implicit-arrow-linebreak': 'off',
     'import/no-anonymous-default-export': 'off',
     'import/no-unassigned-import': 'off',
     'lines-around-comment': 'off',
+    'n/no-sync': 'off',
     'no-async-promise-executor': 'off',
     'no-case-declarations': 'off',
     'no-invalid-this': 'off',
     'no-negated-condition': 'off',
     'no-new': 'off',
     'no-param-reassign': 'off',
+    'no-restricted-syntax': 'off',
     radix: 'off',
     'require-atomic-updates': 'off',
     'jsdoc/match-description': [
-      'error',
+      'off',
       { matchDescription: '^[A-Z`\\d_][\\s\\S]*[.?!`>)}]$' },
     ],
   },