+++ packages/assets-controllers/src/TokensController.test.ts
@@ -1,29 +1,29 @@
-import * as sinon from 'sinon';
-import nock from 'nock';
-import { AddApprovalRequest } from '@metamask/approval-controller';
+import type { AddApprovalRequest } from '@metamask/approval-controller';
 import { ControllerMessenger } from '@metamask/base-controller';
 import contractMaps from '@metamask/contract-metadata';
-import { PreferencesController } from '@metamask/preferences-controller';
 import {
   ApprovalType,
   ChainId,
   ERC20,
   NetworkType,
+  NetworksTicker,
   ORIGIN_METAMASK,
   convertHexToDecimal,
   toHex,
 } from '@metamask/controller-utils';
-import {
+import type {
   NetworkState,
   ProviderConfig,
-  defaultState as defaultNetworkState,
 } from '@metamask/network-controller';
-import {
-  TokensController,
-  TokensControllerMessenger,
-} from './TokensController';
-import { Token } from './TokenRatesController';
+import { defaultState as defaultNetworkState } from '@metamask/network-controller';
+import { PreferencesController } from '@metamask/preferences-controller';
+import nock from 'nock';
+import * as sinon from 'sinon';
+
 import { TOKEN_END_POINT_API } from './token-service';
+import type { Token } from './TokenRatesController';
+import { TokensController } from './TokensController';
+import type { TokensControllerMessenger } from './TokensController';
 
 jest.mock('uuid', () => {
   return {
@@ -40,8 +40,16 @@ const stubCreateEthers = (ctrl: TokensController, res: boolean) => {
   });
 };
 
-const SEPOLIA = { chainId: toHex(11155111), type: NetworkType.sepolia };
-const GOERLI = { chainId: toHex(5), type: NetworkType.goerli };
+const SEPOLIA = {
+  chainId: toHex(11155111),
+  type: NetworkType.sepolia,
+  ticker: NetworksTicker.sepolia,
+};
+const GOERLI = {
+  chainId: toHex(5),
+  type: NetworkType.goerli,
+  ticker: NetworksTicker.goerli,
+};
 
 const controllerName = 'TokensController' as const;