+++ packages/controller-utils/src/siwe.test.ts
@@ -1,4 +1,5 @@
 import { ParsedMessage } from '@spruceid/siwe-parser';
+
 import { detectSIWE, isValidSIWEOrigin } from './siwe';
 
 const mockedParsedMessage = {