+++ packages/controller-utils/src/siwe.ts
@@ -1,5 +1,6 @@
 import { ParsedMessage } from '@spruceid/siwe-parser';
 import { isHexPrefixed } from 'ethereumjs-util';
+
 import { projectLogger, createModuleLogger } from './logger';
 
 const log = createModuleLogger(projectLogger, 'detect-siwe');