+++ packages/message-manager/src/PersonalMessageManager.test.ts
@@ -1,4 +1,6 @@
-import { detectSIWE, SIWEMessage } from '@metamask/controller-utils';
+import type { SIWEMessage } from '@metamask/controller-utils';
+import { detectSIWE } from '@metamask/controller-utils';
+
 import { PersonalMessageManager } from './PersonalMessageManager';
 
 jest.mock('@metamask/controller-utils', () => ({
@@ -81,14 +83,14 @@ describe('PersonalMessageManager', () => {
       messageParams,
       originalRequest,
     );
-    expect(messageId).not.toBeUndefined();
+    expect(messageId).toBeDefined();
     const message = controller.getMessage(messageId);
     if (!message) {
       throw new Error('"message" is falsy');
     }
     expect(message.messageParams.from).toBe(messageParams.from);
     expect(message.messageParams.data).toBe(messageParams.data);
-    expect(message.time).not.toBeUndefined();
+    expect(message.time).toBeDefined();
     expect(message.status).toBe(messageStatus);
     expect(message.type).toBe(messageType);
   });
@@ -123,7 +125,7 @@ describe('PersonalMessageManager', () => {
     };
     await controller.addMessage(firstMessage);
     await controller.addMessage(secondMessage);
-    expect(controller.getUnapprovedMessagesCount()).toStrictEqual(2);
+    expect(controller.getUnapprovedMessagesCount()).toBe(2);
     expect(controller.getUnapprovedMessages()).toStrictEqual({
       [firstMessage.id]: firstMessage,
       [secondMessage.id]: secondMessage,
@@ -142,7 +144,7 @@ describe('PersonalMessageManager', () => {
     if (!message) {
       throw new Error('"message" is falsy');
     }
-    expect(message.status).toStrictEqual('approved');
+    expect(message.status).toBe('approved');
   });
 
   it('should set message status signed', async () => {
@@ -156,7 +158,7 @@ describe('PersonalMessageManager', () => {
       throw new Error('"message" is falsy');
     }
     expect(message.rawSig).toStrictEqual(rawSig);
-    expect(message.status).toStrictEqual('signed');
+    expect(message.status).toBe('signed');
   });
 
   it('should reject message', async () => {
@@ -167,7 +169,7 @@ describe('PersonalMessageManager', () => {
     if (!message) {
       throw new Error('"message" is falsy');
     }
-    expect(message.status).toStrictEqual('rejected');
+    expect(message.status).toBe('rejected');
   });
 
   it('should add message including Ethereum sign in data', async () => {